Add files via upload
This project is for beginners. You can get weather condition on any city of the World. Code is easy to simple to code.
For searching the real data, you need to make api call. so you need to make id on weather api. After activation of api you can use.

Signed-off-by: Bimal BK <109973398+bkbimal250@users.noreply.github.com>